Put importer back

+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/migration/EntityTimestampMigrationV1_46_0Test.java
@@ -206,7 +206,7 @@ class EntityTimestampMigrationV1_46_0Test extends IntegrationTest {
                             transaction.isScheduled(),
                             transaction.getTransactionBytes(),
                             transaction.getTransactionHash(),
-                            transaction.getType().intValue(),
+                            transaction.getType(),
                             transaction.getValidDurationSeconds(),
                             transaction.getValidStartNs());
         }
